Change up how we marshal events and unmarshal them
