+++ events/event.go
@@ -1,57 +1,12 @@
 package events
 
 import (
+	"encoding/json"
 	"fmt"
 	"time"
 
 	"github.com/rs/xid"
 )
-
-//#region EventDataType
-
-// EventObjectType holds the event type we are working.
-type EventObjectType string
-
-const (
-	// SubscriptionDataType describes the type of the data we are holding.
-	// It allow us easily map data from the database to a struct
-	SubscriptionDataType EventObjectType = "subscription"
-
-	// PurchaseDataType describes the type of the data we are holding.
-	// It allow us easily map data from the database to a struct
-	PurchaseDataType EventObjectType = "purchase"
-)
-
-//#endregion EventDataType
-
-//#region event types
-
-// EventType is a unique key defined by event
-type EventType string
-
-const (
-	// AppSubscribedEventType defines the string constant name for the event that happens when
-	// a user subscribes to a plan
-	AppSubscribedEventType EventType = "app.subscription.created"
-
-	// AppSubscriptionCancelledEventType defines the string constant name for the event that happens when
-	// a user cancels a subscription
-	AppSubscriptionCancelledEventType EventType = "app.subscription.cancelled"
-
-	// AppSubscriptionModifiedEventType defines the string constant name for the event that happens when
-	// a subscription is modified
-	AppSubscriptionModifiedEventType EventType = "app.subscription.modified"
-
-	// AppPurchasedEventType indicates which type the struct is. Could make it easier for us data map this event from the database
-	AppPurchasedEventType EventType = "app.purchase.created"
-)
-
-func (et EventType) Valid() bool {
-	return et == AppSubscribedEventType || et == AppPurchasedEventType ||
-		et == AppSubscriptionCancelledEventType || et == AppSubscriptionModifiedEventType
-}
-
-//#endregion event types
 
 // App holds information about the application which the event was triggered
 type App struct {
@@ -116,17 +71,22 @@ type Event struct {
 	When time.Time `bson:"when" json:"when"`
 
 	// Data has the event specific payload
-	Data interface{} `bson:"data" json:"data"`
+	Data json.RawMessage `bson:"data" json:"data"`
 }
 
 // New creates a new event with the status of created
 func New(name EventType, data interface{}) Event {
+	marshalledJson, err := json.Marshal(data)
+	if err != nil {
+		panic(fmt.Sprintf("error creating a new event: %s", err.Error()))
+	}
+
 	return Event{
 		ID:         fmt.Sprintf("evt_%s", xid.New().String()),
 		Type:       name,
 		When:       time.Now(),
 		APIVersion: CurrentAPIVersion,
-		Data:       data,
+		Data:       marshalledJson,
 	}
 }
